refactor(decimal,script,migrations,dbcontext): switch to decimal instead of double for values; add missing category field configuration values for dbcontext and add baseconstraints to precision/scale for both decimals and date (compatibility with date controls that only handle 3 precision decimal timestamps for dates); regen migrations and address the strict check in the seeding that dotnet 10 model builder does when checking context mutations on model creation; include a powershell script to generate local environment variables from .env to allow running migration commands in local environment

+++ src/Services/Expenses/Application/Features/Commands/PatchExpenseCommand.cs
@@ -4,7 +4,7 @@ public sealed record PatchExpenseCommand(
     string Id,
     string Name,
     string Description,
-    double Value,
+    decimal Value,
     DateTime Date,
     string CategoryId,
     string ApplicationUserId
